tidy: move a large function out of an even larger file
